src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiofiles==24.1.0
    # via quart
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via
    #   microsoft-kiota-authentication-azure
    #   quartapp (pyproject.toml)
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   openai
    #   watchfiles
attrs==24.2.0
    # via aiohttp
azure-core==1.31.0
    # via
    #   azure-identity
    #   microsoft-kiota-authentication-azure
azure-identity==1.19.0
    # via quartapp (pyproject.toml)
blinker==1.8.2
    # via
    #   flask
    #   quart
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   flask
    #   quart
    #   uvicorn
cryptography==43.0.1
    # via
    #   azure-identity
    #   msal
    #   pyjwt
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
distro==1.9.0
    # via openai
flask==3.0.3
    # via quart
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gunicorn==23.0.0
    # via quartapp (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   hypercorn
    #   uvicorn
    #   wsproto
h2==4.1.0
    # via
    #   httpx
    #   hypercorn
hpack==4.0.0
    # via h2
httpcore==1.0.6
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx[http2]==0.27.2
    # via
    #   langgraph-sdk
    #   langsmith
    #   microsoft-kiota-http
    #   msgraph-core
    #   openai
httpx-sse==0.4.0
    # via langgraph-sdk
hypercorn==0.17.3
    # via quart
hyperframe==6.0.1
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.4.0
    # via opentelemetry-api
itsdangerous==2.2.0
    # via
    #   flask
    #   quart
jinja2==3.1.4
    # via
    #   flask
    #   quart
jiter==0.6.1
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.3.12
    # via
    #   langchain-openai
    #   langgraph
    #   langgraph-checkpoint
langchain-openai==0.2.2
    # via quartapp (pyproject.toml)
langgraph==0.2.39
    # via quartapp (pyproject.toml)
langgraph-checkpoint==2.0.1
    # via langgraph
langgraph-sdk==0.1.33
    # via langgraph
langsmith==0.1.136
    # via langchain-core
markupsafe==3.0.1
    # via
    #   jinja2
    #   quart
    #   werkzeug
microsoft-kiota-abstractions==1.4.6
    # via
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   msgraph-core
microsoft-kiota-authentication-azure==1.4.6
    # via msgraph-core
microsoft-kiota-http==1.4.6
    # via msgraph-core
msal==1.31.0
    # via
    #   azure-identity
    #   msal-extensions
    #   quartapp (pyproject.toml)
msal-extensions==1.2.0
    # via
    #   azure-identity
    #   quartapp (pyproject.toml)
msgpack==1.1.0
    # via langgraph-checkpoint
msgraph-core==1.1.6
    # via quartapp (pyproject.toml)
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
openai==1.51.2
    # via
    #   langchain-openai
    #   quartapp (pyproject.toml)
opentelemetry-api==1.27.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.27.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
opentelemetry-semantic-conventions==0.48b0
    # via opentelemetry-sdk
orjson==3.10.9
    # via
    #   langgraph-sdk
    #   langsmith
packaging==24.1
    # via
    #   gunicorn
    #   langchain-core
portalocker==2.10.1
    # via msal-extensions
priority==2.0.0
    # via hypercorn
propcache==0.2.0
    # via yarl
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.23.4
    # via pydantic
pyjwt[crypto]==2.9.0
    # via
    #   msal
    #   pyjwt
python-dotenv==1.0.1
    # via
    #   quartapp (pyproject.toml)
    #   uvicorn
pyyaml==6.0.2
    # via
    #   langchain-core
    #   quartapp (pyproject.toml)
    #   uvicorn
quart==0.19.6
    # via quartapp (pyproject.toml)
regex==2024.9.11
    # via tiktoken
requests==2.32.3
    # via
    #   azure-core
    #   langsmith
    #   msal
    #   quartapp (pyproject.toml)
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
six==1.16.0
    # via azure-core
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
std-uritemplate==1.0.6
    # via microsoft-kiota-abstractions
tenacity==9.0.0
    # via langchain-core
tiktoken==0.8.0
    # via langchain-openai
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   azure-core
    #   azure-identity
    #   langchain-core
    #   openai
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
urllib3==2.2.3
    # via
    #   microsoft-kiota-http
    #   requests
uvicorn[standard]==0.31.1
    # via quartapp (pyproject.toml)
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
werkzeug==3.0.4
    # via
    #   flask
    #   quart
    #   quartapp (pyproject.toml)
wrapt==1.16.0
    # via deprecated
wsproto==1.2.0
    # via hypercorn
yarl==1.14.0
    # via aiohttp
zipp==3.20.2
    # via importlib-metadata
